python-services/sms_service/requirements.txt
# FILE PATH: python-services/sms_service/requirements.txt

Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
twilio==8.11.0
APScheduler==3.10.4
